The api logic was split on different files
